refactor: move display logic to display/plaintext.py and display/html.py
